online-electronics-store/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Flask-WTF
WTForms
pymysql
